Use `usize` instead of `u16` for lengths

+++ implementation/sutp/src/chunk.rs
@@ -61,11 +61,11 @@ pub enum CompressionAlgorithm {
 }
 
 const U16_SIZE: usize = mem::size_of::<u16>();
-const U32_SIZE: u16 = mem::size_of::<u32>() as u16;
+const U32_SIZE: usize = mem::size_of::<u32>();
 const ZEROS: [u8; 3] = [0; 3];
 
 /// The constant overhead of serializing a single chunk.
-const BINARY_OVERHEAD: usize = 2 * (U16_SIZE as usize);
+const BINARY_OVERHEAD: usize = 2 * U16_SIZE;
 
 #[allow(dead_code)]
 impl Chunk {
@@ -73,12 +73,11 @@ impl Chunk {
     pub fn binary_len(&self) -> usize {
         BINARY_OVERHEAD + (match self {
             Chunk::Abort | Chunk::Fin | Chunk::Syn => 0,
-            Chunk::CompressionNegotiation(algs) => algs.len() * (U32_SIZE as usize),
+            Chunk::CompressionNegotiation(algs) => algs.len() * U32_SIZE,
             Chunk::Unknown(_, data) | Chunk::Payload(data) =>
                 data.len() + Self::calculate_padding(data.len()),
-            Chunk::Sack(_, list) =>
-                (U32_SIZE as usize) + list.len() * (U32_SIZE as usize),
-            Chunk::SecurityFlag(_) => U32_SIZE as usize,
+            Chunk::Sack(_, list) => U32_SIZE + list.len() * U32_SIZE,
+            Chunk::SecurityFlag(_) => U32_SIZE,
         })
     }
 
@@ -158,14 +157,14 @@ impl Chunk {
     /// chunk list, and not an unexpected EOF. Otherwise returns `Some`.
     pub fn read_from(r: &mut Bytes) -> Result<Option<Self>> {
         let mut buf = r.as_ref().into_buf();
-        if buf.remaining() < (U32_SIZE as usize) {
+        if buf.remaining() < U32_SIZE {
             return Ok(None);
         }
 
         let ty = buf.get_u16_be();
 
         // Advancing the buf doesn't advance the Bytes
-        r.advance(U16_SIZE as usize);
+        r.advance(U16_SIZE);
 
         // Type list as specified at https://laboratory.comsys.rwth-aachen.de/sutp/data-format
         let (variant, bytes_read) = match ty {
@@ -180,7 +179,7 @@ impl Chunk {
         }?;
 
         // Discard padding between chunks
-        let padding = Self::calculate_padding(bytes_read as usize);
+        let padding = Self::calculate_padding(bytes_read);
         if r.len() < padding {
             return Err(io::ErrorKind::UnexpectedEof.into());
         }
@@ -210,7 +209,7 @@ impl Chunk {
         }?;
 
         // Write padding as necessary
-        let padding = Self::calculate_padding(payload_length as usize);
+        let padding = Self::calculate_padding(payload_length);
         w.write_all(&ZEROS[..padding])?;
 
         Ok(())
@@ -225,18 +224,18 @@ impl Chunk {
 // Read implementations
 impl Chunk {
     /// Reads an ABRT chunk.
-    fn read_abrt(r: &mut Bytes) -> Result<(Self, u16)> {
+    fn read_abrt(r: &mut Bytes) -> Result<(Self, usize)> {
         Self::read_flag_chunk(Chunk::Abort, r)
     }
 
     /// Reads an SUTP compression negotiation chunk.
-    fn read_compression_negotiation(r: &mut Bytes) -> Result<(Self, u16)> {
+    fn read_compression_negotiation(r: &mut Bytes) -> Result<(Self, usize)> {
         let mut buf = r.as_ref().into_buf();
-        let len = buf.get_u16_be();
+        let len = buf.get_u16_be() as usize;
 
         // Ensure the list length is valid (i. e. a multiple of size_of::<u32>())
         debug_log_eq!(len % U32_SIZE, 0);
-        assert_size!(buf, len as usize);
+        assert_size!(buf, len);
 
         // You can collect an iterator of results into a result of an iterator
         let list = (0..(len / U32_SIZE))
@@ -246,7 +245,7 @@ impl Chunk {
     }
 
     /// Reads a FIN chunk.
-    fn read_fin(r: &mut Bytes) -> Result<(Self, u16)> {
+    fn read_fin(r: &mut Bytes) -> Result<(Self, usize)> {
         Self::read_flag_chunk(Chunk::Fin, r)
     }
 
@@ -254,7 +253,7 @@ impl Chunk {
     ///
     /// The implementation is based on read_unknown and just
     /// changes the type of the chunk that was read.
-    fn read_payload(r: &mut Bytes) -> Result<(Self, u16)> {
+    fn read_payload(r: &mut Bytes) -> Result<(Self, usize)> {
         Self::read_unknown(0, r)
             .map(|(ch, len)| match ch {
                 Chunk::Unknown(_, data) => (Chunk::Payload(data), len),
@@ -263,13 +262,13 @@ impl Chunk {
     }
 
     /// Reads a SACK chunk.
-    fn read_sack(r: &mut Bytes) -> Result<(Self, u16)> {
+    fn read_sack(r: &mut Bytes) -> Result<(Self, usize)> {
         let mut buf = r.as_ref().into_buf();
-        let len = buf.get_u16_be();
+        let len = buf.get_u16_be() as usize;
 
         debug_log_assert!(len >= U32_SIZE);
         debug_log_eq!(len % U32_SIZE, 0);
-        assert_size!(buf, len as usize);
+        assert_size!(buf, len);
 
         let ack_no = buf.get_u32_be();
         let nak_list = (0..((len - U32_SIZE) / U32_SIZE))
@@ -280,7 +279,7 @@ impl Chunk {
     }
 
     /// Reads a security flag chunk.
-    fn read_security_flag(r: &mut Bytes) -> Result<(Self, u16)> {
+    fn read_security_flag(r: &mut Bytes) -> Result<(Self, usize)> {
         let mut buf = r.as_ref().into_buf();
         let len = buf.get_u16_be();
 
@@ -292,35 +291,35 @@ impl Chunk {
     }
 
     /// Reads a SYN chunk.
-    fn read_syn(r: &mut Bytes) -> Result<(Self, u16)> {
+    fn read_syn(r: &mut Bytes) -> Result<(Self, usize)> {
         Self::read_flag_chunk(Chunk::Syn, r)
     }
 
     /// Reads the data of an unknown chunk into a buffer.
-    fn read_unknown(ty: u16, r: &mut Bytes) -> Result<(Self, u16)> {
+    fn read_unknown(ty: u16, r: &mut Bytes) -> Result<(Self, usize)> {
         let mut buf = r.as_ref().into_buf();
         let len = buf.get_u16_be() as usize;
 
-        r.advance(U16_SIZE as usize);
+        r.advance(U16_SIZE);
         if r.len() < len {
             return Err(io::ErrorKind::UnexpectedEof.into());
         }
 
-        Ok((Chunk::Unknown(ty, r.split_to(len)), len as u16))
+        Ok((Chunk::Unknown(ty, r.split_to(len)), len))
     }
 
     /// Reads a flag (zero-sized) chunk from the reader.
-    fn read_flag_chunk(ch: Chunk, r: &mut Bytes) -> Result<(Self, u16)> {
+    fn read_flag_chunk(ch: Chunk, r: &mut Bytes) -> Result<(Self, usize)> {
         let mut buf = r.as_ref().into_buf();
-        let len = buf.get_u16_be();
+        let len = buf.get_u16_be() as usize;
 
         // Ensure we are given correct data, but otherwise discard what we've been
         // given for a robust implementation.
         debug_log_eq!(len, 0);
-        assert_size!(buf, len as usize);
+        assert_size!(buf, len);
 
         // Reading from the buffer doesn't advance the Bytes
-        r.advance(U16_SIZE + len as usize);
+        r.advance(U16_SIZE + len);
 
         Ok((ch, len))
     }
@@ -329,7 +328,7 @@ impl Chunk {
 // Write implementations
 impl Chunk {
     /// Writes an ABRT chunk to the given writer.
-    fn write_abrt(w: &mut impl Write) -> Result<u16> {
+    fn write_abrt(w: &mut impl Write) -> Result<usize> {
         Self::write_chunk_header(0x3, 0, w)?;
         Ok(0)
     }
@@ -338,8 +337,8 @@ impl Chunk {
     fn write_compression_negotiation(
         list: &[CompressionAlgorithm],
         w: &mut impl Write,
-    ) -> Result<u16> {
-        let payload_size = list.len() * (U32_SIZE as usize);
+    ) -> Result<usize> {
+        let payload_size = list.len() * U32_SIZE;
         assert!(payload_size <= u16::MAX as usize);
 
         Self::write_chunk_header(0xa0, payload_size as u16, w)?;
@@ -348,23 +347,23 @@ impl Chunk {
             w.write_u32::<NetworkEndian>(alg.into())?;
         }
 
-        Ok(payload_size as u16)
+        Ok(payload_size)
     }
 
     /// Writes a FIN chunk to the given writer.
-    fn write_fin(w: &mut impl Write) -> Result<u16> {
+    fn write_fin(w: &mut impl Write) -> Result<usize> {
         Self::write_chunk_header(0x2, 0, w)?;
         Ok(0)
     }
 
     /// Writes a payload chunk containing the given data to the given writer.
-    fn write_payload(data: &[u8], w: &mut impl Write) -> Result<u16> {
+    fn write_payload(data: &[u8], w: &mut impl Write) -> Result<usize> {
         assert!(data.len() <= u16::MAX as usize);
 
         Self::write_chunk_header(0x0, data.len() as u16, w)?;
         w.write_all(&data)?;
 
-        Ok(data.len() as u16)
+        Ok(data.len())
     }
 
     /// Writes a SACK chunk containing the given data to the given writer.
@@ -372,9 +371,8 @@ impl Chunk {
         ack_no: u32,
         nak_list: &[u32],
         w: &mut impl Write,
-    ) -> Result<u16> {
-        let payload_len = U32_SIZE as usize +
-            nak_list.len() as usize * U32_SIZE as usize;
+    ) -> Result<usize> {
+        let payload_len = U32_SIZE + nak_list.len() * U32_SIZE;
 
         assert!(payload_len <= u16::MAX as usize);
         Self::write_chunk_header(0x4, payload_len as u16, w)?;
@@ -385,11 +383,11 @@ impl Chunk {
             w.write_u32::<NetworkEndian>(nak_no)?;
         }
 
-        Ok(payload_len as u16)
+        Ok(payload_len)
     }
 
     /// Writes the security flag chunk to the given writer.
-    fn write_security_flag(is_insecure: bool, w: &mut impl Write) -> Result<u16> {
+    fn write_security_flag(is_insecure: bool, w: &mut impl Write) -> Result<usize> {
         Self::write_chunk_header(0xfe, 1, w)?;
         w.write_u8(if is_insecure { 1 } else { 0 })?;
 
@@ -397,19 +395,19 @@ impl Chunk {
     }
 
     /// Writes a SYN to the given writer.
-    fn write_syn(w: &mut impl Write) -> Result<u16> {
+    fn write_syn(w: &mut impl Write) -> Result<usize> {
         Self::write_chunk_header(0x1, 0, w)?;
         Ok(0)
     }
 
     /// Writes an unknown chunk of the given type.
-    fn write_unknown(ty: u16, data: &[u8], w: &mut impl Write) -> Result<u16> {
+    fn write_unknown(ty: u16, data: &[u8], w: &mut impl Write) -> Result<usize> {
         assert!(data.len() <= u16::MAX as usize);
 
         Self::write_chunk_header(ty, data.len() as u16, w)?;
         w.write_all(data)?;
 
-        Ok(data.len() as u16)
+        Ok(data.len())
     }
 
     /// Writes chunk type and payload length to the given writer.
